Refactor new(Batch)ApiRequest into api pkg
